Try and fix build on mac.
